@@ -1,45 +1,46 @@
 name: unicaen_timetable
 description: Allows the students of the University of Caen Normandie to check their timetable directly from their smartphone.
-version: 4.0.4+93
+version: 5.0.0+94
 
 environment:
-  sdk: '>=3.0.0 <4.0.0'
+  sdk: '>=3.6.0 <4.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
+  flutter_localizations:
+    sdk: flutter
   flutter_week_view: ^1.3.0
-  jovial_svg: ^1.1.23
-  ez_localization: ^1.1.1
+  jovial_svg: ^1.1.25
   intl: any
   path_provider: ^2.1.5
-  http: ^1.2.2
+  http: ^1.3.0
   background_fetch: ^1.3.7
   flutter_colorpicker: ^1.1.0
-  share_plus: ^10.1.1
-  rate_my_app: ^2.2.0
+  share_plus: ^10.1.4
+  rate_my_app: ^2.3.1
   url_launcher: ^6.3.1
-  package_info_plus: ^8.1.0
-  device_info_plus: ^11.1.0
+  package_info_plus: ^8.2.1
+  device_info_plus: ^11.3.0
   scrollable_positioned_list: ^0.3.8
-  google_mobile_ads: ^5.2.0
   flutter_riverpod: ^2.6.1
   flutter_widget_from_html_core: ^0.15.2
-  shared_preferences: ^2.3.2
-  firebase_crashlytics: ^4.1.3
+  shared_preferences: ^2.5.2
+  firebase_crashlytics: ^4.3.2
+  drift: ^2.25.0
+  drift_flutter: ^0.2.4
+  slang: ^4.4.1
+  slang_flutter: ^4.4.0
+  firebase_core: ^3.11.0
 
 dev_dependencies:
   flutter_lints: ^5.0.0
+  drift_dev: ^2.25.1
+  build_runner: ^2.4.15
 
 flutter:
   uses-material-design: true
   assets:
     - 'assets/'
-    - 'assets/languages/'
     - 'assets/intro/'
-    - 'assets/fonts/'
     - 'assets/about/'
-  fonts:
-    - family: MaterialCommunityIcons
-      fonts:
-        - asset: assets/fonts/MaterialCommunityIcons.ttf
