@@ -1,26 +1,48 @@
 import 'dart:math' as math;
 
-import 'package:ez_localization/ez_localization.dart';
 import 'package:flutter/material.dart' hide Page;
-import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_widget_from_html_core/flutter_widget_from_html_core.dart';
+import 'package:unicaen_timetable/i18n/translations.g.dart';
 import 'package:unicaen_timetable/pages/page.dart';
+import 'package:unicaen_timetable/widgets/drawer/list_title.dart';
 
-/// A page that allows the user to contact me in case of any bug occurred / improvements needed.
-class BugsImprovementsPage extends Page {
-  /// The page identifier.
-  static const String id = 'bugs_improvements';
-
-  /// Creates a new bugs / improvements page instance.
-  const BugsImprovementsPage({
+/// The bugs / improvements page list tile.
+class BugsImprovementsPageListTile extends StatelessWidget {
+  /// Creates a new bugs / improvements page list tile.
+  const BugsImprovementsPageListTile({
     super.key,
-  }) : super(
-          pageId: id,
-          icon: Icons.bug_report,
-        );
+  });
 
   @override
-  Widget build(BuildContext context, WidgetRef ref) => Center(
+  Widget build(BuildContext context) => PageListTitle(
+        page: BugsImprovementsPage(),
+        title: translations.bugsImprovements.title,
+        icon: Icons.bug_report,
+      );
+}
+
+/// The bugs / improvements page app bar.
+class BugsImprovementsPageAppBar extends StatelessWidget {
+  /// Creates a new about page app bar.
+  const BugsImprovementsPageAppBar({
+    super.key,
+  });
+
+  @override
+  Widget build(BuildContext context) => AppBar(
+        title: Text(translations.about.title),
+      );
+}
+
+/// The bugs / improvements page widget.
+class BugsImprovementsPageWidget extends StatelessWidget {
+  /// Creates a new bugs / improvements page instance.
+  const BugsImprovementsPageWidget({
+    super.key,
+  });
+
+  @override
+  Widget build(BuildContext context) => Center(
         child: SingleChildScrollView(
           padding: const EdgeInsets.all(40),
           child: Column(
@@ -33,8 +55,8 @@ class BugsImprovementsPage extends Page {
                   color: Colors.red[400],
                 ),
               ),
-              HtmlWidget(context.getString('bugs_improvements.message.github')),
-              HtmlWidget(context.getString('bugs_improvements.message.website')),
+              HtmlWidget(translations.bugsImprovements.message.github),
+              HtmlWidget(translations.bugsImprovements.message.website),
             ],
           ),
         ),
