@@ -1,11 +1,8 @@
-import 'package:flutter/material.dart' hide Page;
+import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:jovial_svg/jovial_svg.dart';
-import 'package:unicaen_timetable/model/lessons/user/repository.dart';
-import 'package:unicaen_timetable/model/lessons/user/user.dart';
-import 'package:unicaen_timetable/model/settings/settings.dart';
-import 'package:unicaen_timetable/theme.dart';
+import 'package:unicaen_timetable/model/user/user.dart';
 
 /// The drawer header.
 class DrawerHeader extends ConsumerWidget {
@@ -16,20 +13,23 @@ class DrawerHeader extends ConsumerWidget {
 
   @override
   Widget build(BuildContext context, WidgetRef ref) {
-    UserRepository userRepository = ref.watch(userRepositoryProvider);
-    User? user = userRepository.user;
+    User? user = ref.watch(userProvider).valueOrNull;
     if (user == null) {
       return const SizedBox.shrink();
     }
 
-    UnicaenTimetableTheme theme = ref.watch(settingsModelProvider).resolveTheme(context);
+    String username = user.username;
+    if (username.endsWith('@etu.unicaen.fr')) {
+      username = username.split('@').first;
+    }
+    String email = '$username@etu.unicaen.fr';
     return UserAccountsDrawerHeader(
-      accountName: Text(user.usernameWithoutAt),
-      accountEmail: Text(user.username.contains('@') ? user.username : ('${user.username}@etu.unicaen.fr')),
+      accountName: Text(username),
+      accountEmail: Text(email),
       currentAccountPicture: ScalableImageWidget.fromSISource(
         si: ScalableImageSource.fromSI(rootBundle, 'assets/icon.si'),
       ),
-      decoration: BoxDecoration(color: theme.actionBarColor),
+      decoration: BoxDecoration(color: Theme.of(context).appBarTheme.backgroundColor),
     );
   }
 }
